Make MirroringSession reusable
This patch moves common code that is used to implement MirroringSession
in both psa_switch and simple_switch into the BaseSwitch class. This
change allows to decouple pi_clone_imp from simple_switch and make it
reusable.
