+++ targets/simple_switch/simple_switch.h
@@ -85,14 +85,6 @@ class SimpleSwitch : public BaseSwitch {
 
   void swap_notify_() override;
 
-  bool mirroring_add_session(mirror_id_t mirror_id,
-                             const MirroringSessionConfig &config);
-
-  bool mirroring_delete_session(mirror_id_t mirror_id);
-
-  bool mirroring_get_session(mirror_id_t mirror_id,
-                             MirroringSessionConfig *config) const;
-
   int set_egress_priority_queue_depth(size_t port, size_t priority,
                                       const size_t depth_pkts);
   int set_egress_queue_depth(size_t port, const size_t depth_pkts);
@@ -120,8 +112,6 @@ class SimpleSwitch : public BaseSwitch {
 
  private:
   static constexpr size_t nb_egress_threads = 4u;
-
-  class MirroringSessions;
 
   class InputBuffer;
 
@@ -178,7 +168,6 @@ class SimpleSwitch : public BaseSwitch {
   std::shared_ptr<McSimplePreLAG> pre;
   clock::time_point start;
   bool with_queueing_metadata{false};
-  std::unique_ptr<MirroringSessions> mirroring_sessions;
 };
 
 #endif  // SIMPLE_SWITCH_SIMPLE_SWITCH_H_
